+++ src/Internals/Commands/ProgressBar/ProgressBarRegistrationMessage.php
@@ -3,10 +3,10 @@
 namespace HDSSolutions\Console\Parallel\Internals\Commands\ProgressBar;
 
 use HDSSolutions\Console\Parallel\Internals\Commands\ParallelCommandMessage;
-use HDSSolutions\Console\Parallel\Internals\ProgressBarWorker;
+use HDSSolutions\Console\Parallel\Internals\Runner;
 
 /**
- * Message sent to {@see ProgressBarWorker} to execute {@see ProgressBarWorker::registerWorker()} action
+ * Message sent to {@see Runner} to execute {@see Runner\HasSharedProgressBar::registerProgressBar()} action
  */
 final readonly class ProgressBarRegistrationMessage extends ParallelCommandMessage {
 
@@ -15,7 +15,7 @@ final readonly class ProgressBarRegistrationMessage extends ParallelCommandMessa
      * @param  int  $steps
      */
     public function __construct(string $worker, int $steps = 0) {
-        parent::__construct('register_worker', [ $worker, $steps ]);
+        parent::__construct('register_progress_bar', [ $worker, $steps ]);
     }
 
 }